refactor(core): 🔨 add Image partial implementations
