+++ Sources/ImagePlayground.Core/Image.cs
@@ -1,15 +1,10 @@
 ﻿using System;
 using System.IO;
-using System.Numerics;
 using Codeuctivity.ImageSharpCompare;
 using SixLabors.ImageSharp;
-using SixLabors.ImageSharp.Drawing;
-using SixLabors.ImageSharp.Drawing.Processing;
 using SixLabors.ImageSharp.Formats;
 using SixLabors.ImageSharp.Metadata;
 using SixLabors.ImageSharp.PixelFormats;
-using SixLabors.ImageSharp.Processing;
-using SixLabors.ImageSharp.Processing.Extensions.Transforms;
 
 namespace ImagePlayground;
 
@@ -37,67 +32,6 @@ public partial class Image : IDisposable {
 
     /// <summary>Gets the frame collection for multi-frame images.</summary>
     public ImageFrameCollection Frames => _image.Frames;
-
-
-
-    /// <summary>
-    /// Applies an adaptive threshold filter to the current image.
-    /// </summary>
-    public void AdaptiveThreshold() {
-        _image.Mutate(x => x.AdaptiveThreshold());
-    }
-
-    /// <summary>
-    /// Automatically rotates the image according to EXIF orientation data.
-    /// </summary>
-    public void AutoOrient() {
-        _image.Mutate(x => x.AutoOrient());
-    }
-
-    /// <summary>
-    /// Fills the background with the specified <paramref name="color"/>.
-    /// </summary>
-    /// <param name="color">Fill color.</param>
-    public void BackgroundColor(Color color) {
-        _image.Mutate(x => x.BackgroundColor(color));
-    }
-
-    /// <summary>
-    /// Converts the image to black and white.
-    /// </summary>
-    public void BlackWhite() {
-        _image.Mutate(x => x.BlackWhite());
-    }
-
-    /// <summary>
-    /// Adjusts brightness by the specified <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Brightness adjustment factor.</param>
-    public void Brightness(float amount) {
-        _image.Mutate(x => x.Brightness(amount));
-    }
-
-    /// <summary>
-    /// Applies a bokeh blur effect.
-    /// </summary>
-    public void BokehBlur() {
-        _image.Mutate(x => x.BokehBlur());
-    }
-
-    /// <summary>
-    /// Applies a simple box blur.
-    /// </summary>
-    public void BoxBlur() {
-        _image.Mutate(x => x.BoxBlur());
-    }
-
-    /// <summary>
-    /// Adjusts contrast by the specified <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Contrast adjustment factor.</param>
-    public void Contrast(float amount) {
-        _image.Mutate(x => x.Contrast(amount));
-    }
 
     /// <summary>
     /// Compares the current image with <paramref name="imageToCompare"/>.
@@ -153,318 +87,6 @@ public partial class Image : IDisposable {
                 }
             }
         }
-    }
-
-    /// <summary>
-    /// Crops the image to the specified <paramref name="rectangle"/>.
-    /// </summary>
-    /// <param name="rectangle">Crop rectangle.</param>
-    public void Crop(Rectangle rectangle) {
-        _image.Mutate(x => x.Crop(rectangle));
-    }
-
-    /// <summary>
-    /// Applies a dithering effect.
-    /// </summary>
-    public void Dither() {
-        _image.Mutate(x => x.Dither());
-    }
-
-    /// <summary>
-    /// Draws a line ending at <paramref name="pointF"/> using the specified <paramref name="color"/> and <paramref name="thickness"/>.
-    /// </summary>
-    /// <param name="color">Line color.</param>
-    /// <param name="thickness">Line thickness.</param>
-    /// <param name="pointF">End point of the line.</param>
-    public void DrawLines(Color color, float thickness, PointF pointF) {
-        _image.Mutate(x => x.DrawLine(color, thickness, pointF));
-    }
-
-    /// <summary>
-    /// Draws a polygon with the specified parameters.
-    /// </summary>
-    /// <param name="color">Polygon color.</param>
-    /// <param name="thickness">Outline thickness.</param>
-    /// <param name="pointF">Polygon vertex.</param>
-    public void DrawPolygon(Color color, float thickness, PointF pointF) {
-        _image.Mutate(x => x.DrawPolygon(color, thickness, pointF));
-    }
-
-    /// <summary>
-    /// Applies a color <paramref name="colorMatrix"/> filter.
-    /// </summary>
-    /// <param name="colorMatrix">Color matrix to apply.</param>
-    public void Filter(ColorMatrix colorMatrix) {
-        _image.Mutate(x => x.Filter(colorMatrix));
-    }
-
-    /// <summary>
-    /// Fills the image with <paramref name="color"/>.
-    /// </summary>
-    /// <param name="color">Fill color.</param>
-    public void Fill(Color color) {
-        _image.Mutate(x => x.Fill(color));
-    }
-
-    /// <summary>
-    /// Fills the specified <paramref name="rectangle"/> with <paramref name="color"/>.
-    /// </summary>
-    /// <param name="color">Fill color.</param>
-    /// <param name="rectangle">Target rectangle.</param>
-    public void Fill(Color color, Rectangle rectangle) {
-        _image.Mutate(x => x.Fill(color, rectangle));
-    }
-
-    /// <summary>
-    /// Flips the image using the provided <paramref name="flipMode"/>.
-    /// </summary>
-    /// <param name="flipMode">Flip mode.</param>
-    public void Flip(FlipMode flipMode) {
-        _image.Mutate(x => x.Flip(flipMode));
-    }
-
-    /// <summary>
-    /// Applies a Gaussian blur using the optional <paramref name="sigma"/> value.
-    /// </summary>
-    /// <param name="sigma">Blur radius.</param>
-    public void GaussianBlur(float? sigma) {
-        if (sigma != null) {
-            _image.Mutate(x => x.GaussianBlur(sigma.Value));
-        } else {
-            _image.Mutate(x => x.GaussianBlur());
-        }
-    }
-
-    /// <summary>
-    /// Sharpens the image using a Gaussian algorithm.
-    /// </summary>
-    /// <param name="sigma">Sharpen strength.</param>
-    public void GaussianSharpen(float? sigma) {
-        if (sigma != null) {
-            _image.Mutate(x => x.GaussianSharpen(sigma.Value));
-        } else {
-            _image.Mutate(x => x.GaussianSharpen());
-        }
-    }
-
-    /// <summary>
-    /// Performs histogram equalization on the image.
-    /// </summary>
-    public void HistogramEqualization() {
-        _image.Mutate(x => x.HistogramEqualization());
-    }
-
-    /// <summary>
-    /// Shifts hue by the specified <paramref name="degrees"/>.
-    /// </summary>
-    /// <param name="degrees">Hue rotation in degrees.</param>
-    public void Hue(float degrees) {
-        _image.Mutate(x => x.Hue(degrees));
-    }
-
-    /// <summary>
-    /// Converts the image to grayscale using the specified <paramref name="grayscaleMode"/>.
-    /// </summary>
-    /// <param name="grayscaleMode">Grayscale conversion mode.</param>
-    public void Grayscale(GrayscaleMode grayscaleMode = GrayscaleMode.Bt709) {
-        _image.Mutate(x => x.Grayscale(grayscaleMode));
-    }
-
-    /// <summary>
-    /// Applies a Kodachrome color filter.
-    /// </summary>
-    public void Kodachrome() {
-        _image.Mutate(x => x.Kodachrome());
-    }
-
-    /// <summary>
-    /// Adjusts lightness by <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Lightness factor.</param>
-    public void Lightness(float amount) {
-        _image.Mutate(x => x.Lightness(amount));
-    }
-
-    /// <summary>
-    /// Applies a lomograph effect.
-    /// </summary>
-    public void Lomograph() {
-        _image.Mutate(x => x.Lomograph());
-    }
-
-    /// <summary>
-    /// Inverts the colors of the image.
-    /// </summary>
-    public void Invert() {
-        _image.Mutate(x => x.Invert());
-    }
-
-    /// <summary>
-    /// Changes opacity by the given <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Opacity factor.</param>
-    public void Opacity(float amount) {
-        _image.Mutate(x => x.Opacity(amount));
-    }
-
-    /// <summary>
-    /// Applies a Polaroid style filter.
-    /// </summary>
-    public void Polaroid() {
-        _image.Mutate(x => x.Polaroid());
-    }
-
-    /// <summary>
-    /// Pixelates the image with a default size.
-    /// </summary>
-    public void Pixelate() {
-        _image.Mutate(x => x.Pixelate());
-    }
-
-    /// <summary>
-    /// Pixelates the image using the specified <paramref name="size"/>.
-    /// </summary>
-    /// <param name="size">Pixel block size.</param>
-    public void Pixelate(int size) {
-        _image.Mutate(x => x.Pixelate(size));
-    }
-
-    /// <summary>
-    /// Applies an oil paint effect using default options.
-    /// </summary>
-    public void OilPaint() {
-        _image.Mutate(x => x.OilPaint());
-    }
-
-    /// <summary>
-    /// Applies an oil paint effect with the given parameters.
-    /// </summary>
-    /// <param name="levels">Number of intensity levels.</param>
-    /// <param name="brushSize">Brush size.</param>
-    public void OilPaint(int levels, int brushSize) {
-        _image.Mutate(x => x.OilPaint(levels, brushSize));
-    }
-
-    /// <summary>
-    /// Rotates the image using the specified <paramref name="rotateMode"/>.
-    /// </summary>
-    /// <param name="rotateMode">Rotation mode.</param>
-    public void Rotate(RotateMode rotateMode) {
-        _image.Mutate(x => x.Rotate(rotateMode));
-    }
-
-    /// <summary>
-    /// Rotates the image by an arbitrary number of <paramref name="degrees"/>.
-    /// </summary>
-    /// <param name="degrees">Angle in degrees.</param>
-    public void Rotate(float degrees) {
-        _image.Mutate(x => x.Rotate(degrees: degrees));
-    }
-
-    /// <summary>
-    /// Rotates and flips the image in a single operation.
-    /// </summary>
-    /// <param name="rotateMode">Rotation mode.</param>
-    /// <param name="flipMode">Flip mode.</param>
-    public void RotateFlip(RotateMode rotateMode, FlipMode flipMode) {
-        _image.Mutate(x => x.RotateFlip(rotateMode, flipMode));
-    }
-
-    /// <summary>
-    /// Resizes the image to the specified dimensions.
-    /// </summary>
-    /// <param name="width">New width.</param>
-    /// <param name="height">New height.</param>
-    /// <param name="keepAspectRatio">Maintain aspect ratio if possible.</param>
-    public void Resize(int? width, int? height, bool keepAspectRatio = true) {
-        if (width == null && height == null) {
-            return;
-        }
-
-        var options = new ResizeOptions();
-        if (keepAspectRatio && (width == null || height == null)) {
-            options.Mode = ResizeMode.Max;
-            if (width == null) {
-                int calculatedWidth = (int)Math.Round(height!.Value * _image.Width / (double)_image.Height);
-                options.Size = new Size(calculatedWidth, height.Value);
-            } else {
-                int calculatedHeight = (int)Math.Round(width.Value * _image.Height / (double)_image.Width);
-                options.Size = new Size(width.Value, calculatedHeight);
-            }
-        } else if (keepAspectRatio) {
-            options.Mode = ResizeMode.Max;
-            options.Size = new Size(width ?? _image.Width, height ?? _image.Height);
-        } else {
-            options.Mode = ResizeMode.Stretch;
-            options.Size = new Size(width ?? _image.Width, height ?? _image.Height);
-        }
-
-        _image.Mutate(x => x.Resize(options));
-    }
-
-    /// <summary>
-    /// Resizes the image by a <paramref name="percentage"/> of the current size.
-    /// </summary>
-    /// <param name="percentage">Scale percentage.</param>
-    public void Resize(int percentage) {
-        if (percentage <= 0) {
-            throw new ArgumentOutOfRangeException(nameof(percentage));
-        }
-
-        int width = _image.Width * percentage / 100;
-        int height = _image.Height * percentage / 100;
-        var options = new ResizeOptions {
-            Mode = ResizeMode.Stretch,
-            Size = new Size(width, height)
-        };
-        _image.Mutate(x => x.Resize(options));
-    }
-
-    /// <summary>
-    /// Changes saturation by the specified <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Saturation factor.</param>
-    public void Saturate(float amount) {
-        _image.Mutate(x => x.Saturate(amount));
-    }
-
-    /// <summary>
-    /// Applies a sepia tone effect.
-    /// </summary>
-    public void Sepia() {
-        _image.Mutate(x => x.Sepia());
-    }
-
-    /// <summary>
-    /// Applies a sepia tone effect with the given <paramref name="amount"/>.
-    /// </summary>
-    /// <param name="amount">Sepia intensity.</param>
-    public void Sepia(float amount) {
-        _image.Mutate(x => x.Sepia(amount));
-    }
-
-    /// <summary>
-    /// Skews the image by the specified angles.
-    /// </summary>
-    /// <param name="degreesX">Skew angle on the X axis.</param>
-    /// <param name="degreesY">Skew angle on the Y axis.</param>
-    public void Skew(float degreesX, float degreesY) {
-        _image.Mutate(x => x.Skew(degreesX, degreesY));
-    }
-
-    /// <summary>
-    /// Adds a vignette effect using default options.
-    /// </summary>
-    public void Vignette() {
-        _image.Mutate(x => x.Vignette());
-    }
-
-    /// <summary>
-    /// Adds a vignette effect using the specified <paramref name="color"/>.
-    /// </summary>
-    /// <param name="color">Vignette color.</param>
-    public void Vignette(Color color) {
-        _image.Mutate(x => x.Vignette(color));
     }
 
     /// <summary>
