+++ Sources/ImagePlayground.Core/ImagePlayground.Core.csproj
@@ -22,6 +22,13 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Compile Remove="Image.Filters.cs;Image.Drawing.cs;Image.Transforms.cs" />
+    <Compile Include="Image.Filters.cs" />
+    <Compile Include="Image.Drawing.cs" />
+    <Compile Include="Image.Transforms.cs" />
+  </ItemGroup>
+
+  <ItemGroup>
     <Using Include="System.Collections" />
     <Using Include="System.Threading.Tasks" />
     <Using Include="System.Collections.Concurrent" />
